+++ src/main/java/com/github/correa/finalreality/gui/FinalReality.java
@@ -12,7 +12,7 @@ import javafx.stage.Stage;
  * <Complete here with the details of the implemented application>
  *
  * @author Ignacio Slater Muñoz.
- * @author <Your name>
+ * @author Benjamín Correa Karstulovic
  */
 public class FinalReality extends Application {